Move swimlane implementation to it's own file
