+++ scripts/apps/search/components/Item.tsx
@@ -22,8 +22,7 @@ import {
     ListItemInfo,
 } from './index';
 import {closeActionsMenu} from '../helpers';
-import {DEFAULT_SWIMLANE_FIELDS_CONFIG} from '../constants';
-import {SwimlaneField} from './SwimlaneField';
+import {ItemSwimlane} from './ItemSwimlane';
 
 const actionsMenuDefaultTemplate = (toggle, stopEvent) => (
     <div className="item-right toolbox">
@@ -189,77 +188,14 @@ export class Item extends React.Component<any, any> {
 
         const getTemplate = () => {
             if (this.props.view === 'swimlane2') {
-                const swimlaneViewFieldsConfig = get(
-                    this.props.svc.config, 'swimlaneViewFields',
-                    DEFAULT_SWIMLANE_FIELDS_CONFIG,
-                );
-
-                const renderGroup = (groups) => {
-                    return groups.map((group, groupIndex) => (
-                        <span
-                            key={groupIndex}
-                            className={classNames({
-                                'sd-list-item__column': true,
-                                'sd-list-item__column--grow': group.ellipsis === true,
-                                'sd-list-item__column--no-border': groupIndex === groups.length - 1,
-                            })}>
-                            <span className="sd-list-item__row">
-                                <span className={classNames({'sd-overflow-ellipsis': group.ellipsis === true})}>
-                                    {
-                                        group.fields
-                                            .map((field, fieldIndex) => (
-                                                <SwimlaneField
-                                                    key={fieldIndex}
-                                                    fieldId={field}
-                                                    item={item}
-                                                    svc={this.props.svc}
-                                                />
-                                            ))
-                                    }
-                                </span>
-                            </span>
-                        </span>
-                    ));
-                };
-
                 return (
-                    <div className="sd-list-item" style={{borderBottom: '1px solid #e7e7e7'}}>
-                        <span
-                            style={{
-                                minWidth: '4px',
-                                maxWidth: '4px',
-                                background: isLocked ? '#e51c23' : 'transparent',
-                            }}>
-                        </span>
-                        {
-                            item.archiveError
-                                ? <span className="sd-list-item__column"><ErrorBox svc={this.props.svc} /></span>
-                                : null
-                        }
-                        <span className="sd-list-item__column">
-                            <ListTypeIcon
-                                item={item}
-                                onMultiSelect={this.props.onMultiSelect}
-                                swimlane={this.props.swimlane}
-                                svc={this.props.svc}
-                            />
-                        </span>
-                        {renderGroup(swimlaneViewFieldsConfig.left)}
-                        <span className="sd-list-item--element-grow"></span>
-                        {renderGroup(swimlaneViewFieldsConfig.right)}
-                        {
-                            getActionsMenu((toggle, stopEvent) => (
-                                <div className="sd-list-item__action-menu">
-                                    <button
-                                        className="icn-btn"
-                                        onClick={toggle}
-                                        onDoubleClick={stopEvent}>
-                                        <i className="icon-dots-vertical"></i>
-                                    </button>
-                                </div>
-                            ))
-                        }
-                    </div>
+                    <ItemSwimlane
+                        item={item}
+                        isLocked={isLocked}
+                        getActionsMenu={getActionsMenu}
+                        onMultiSelect={this.props.onMultiSelect}
+                        svc={this.props.svc}
+                    />
                 );
             } else if (this.props.view === 'mgrid') {
                 return (
@@ -317,7 +253,6 @@ export class Item extends React.Component<any, any> {
                         <ListTypeIcon
                             item={item}
                             onMultiSelect={this.props.onMultiSelect}
-                            swimlane={this.props.swimlane}
                             svc={this.props.svc}
                         />
                         {
